Bulk upload added

+++ src/main/java/com/api/arv/model/document/ProductItem.java
@@ -23,9 +23,9 @@ public class ProductItem {
     private String subCategoryId;
     private String subSubCategoryId;
     private String productVariation;
-    private Integer mrpPrice;
-    private Integer sellingPrice;
+    private Double mrpPrice;
+    private Double sellingPrice;
     private Integer quantity;
     private Integer stock;
-    private Integer discount;
+    private Double discount;
 }
